+++ .storybook/preview.tsx
@@ -105,6 +105,18 @@ const preview: Preview = {
           color: "#fff",
           description: "TEDI-ready",
         },
+        partiallyTediReady: {
+          background: '#9bbb5f',
+          color: '#fff',
+          description:
+            'This component lacks some TEDI-Ready functionality, e.g it may rely on another component that has not yet been developed',
+        },
+        mobileViewDifference: {
+          background: '#99BDDA',
+          color: '#000',
+          description:
+            'This component has a different layout on mobile. Use the mobile breakpoint or resize the browser window to review the mobile design.',
+        },
       },
     },
   },